Re #45 Sync arm and gripper trajectory playback
In RSDK 1.1.1, the amount of time required to move to the initial
position for trajectory playback was not taken into account when
calculating when to start gripper playback. By subtracting this
amount of time from gripper playback start, we end up with
synchronized (or nearly) gripper movements with respect to the
robot's arm movements.